api-sgd/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
dotenv
pymysql
mysql-connector-python
patentes-vehiculares-chile
rut-chile